# Conflicts:
#	www/Themes/Centreon-2/style.css

+++ www/include/configuration/configObject/service/DB-Func.php
@@ -722,7 +722,7 @@ function updateServiceInDB ($service_id = null, $from_MC = false, $params = arra
     // 2 - MC with addition of new host/hg parent
     // 3 - Normal update
     if (isset($ret["mc_mod_Pars"]["mc_mod_Pars"]) && $ret["mc_mod_Pars"]["mc_mod_Pars"]) {
-        updateServiceHost($service_id, $params);
+        updateServiceHost($service_id, $params, true);
     } elseif (isset($ret["mc_mod_Pars"]["mc_mod_Pars"]) && !$ret["mc_mod_Pars"]["mc_mod_Pars"]) {
         updateServiceHost_MC($service_id);
     } else {
@@ -1226,17 +1226,17 @@ function updateService($service_id = null, $from_MC = false, $params = array())
     /*
      * Update demand macros
      */
-    $macroDescription = array();
-    foreach ($_REQUEST as $nam => $ele ) {
-        if (preg_match_all("/^macroDescription_(\w+)$/", $nam, $matches, PREG_SET_ORDER)) {
-            foreach ($matches as $match) {
-                $macroDescription[$match[1]] = $ele;
+    if (isset($_REQUEST['macroInput']) && isset($_REQUEST['macroValue'])) {
+        $macroDescription = array();
+        foreach ($_REQUEST as $nam => $ele ) {
+            if (preg_match_all("/^macroDescription_(\w+)$/", $nam, $matches, PREG_SET_ORDER)) {
+                foreach ($matches as $match) {
+                    $macroDescription[$match[1]] = $ele;
+                }
             }
         }
-    }
- 
-    if (isset($_REQUEST['macroInput']) && isset($_REQUEST['macroValue'])) {
-        $service->insertMacro($service_id, $_REQUEST['macroInput'], $_REQUEST['macroValue'], (!isset($_REQUEST['macroPassword']) ? 0 : $_REQUEST['macroPassword']), $macroDescription,false, $ret["command_command_id"]);
+        $service->insertMacro($service_id, $_REQUEST['macroInput'], $_REQUEST['macroValue'], (!isset($_REQUEST['macroPassword']) ? 0 : $_REQUEST['macroPassword']),
+            $macroDescription,$from_MC, $ret["command_command_id"]);
     } else {
         $pearDB->query("DELETE FROM on_demand_macro_service WHERE svc_svc_id = '".CentreonDB::escape($service_id)."'");
     }
@@ -1528,7 +1528,9 @@ function updateService_MC($service_id = null, $params = array())
                               $_REQUEST['macroValue'],
                               $_REQUEST['macroPassword'],
                               $macroDescription,
-                              true
+                              true,
+                              false,
+                              $_REQUEST['macroFrom']
                               );
     }
     if (isset($ret['criticality_id']) && $ret['criticality_id']) {
@@ -1953,7 +1955,7 @@ function updateServiceHostContactsInheritance($service_id = null, $ret = array()
     $DBRESULT =& $pearDB->query($rq);
 }
     
-function updateServiceHost($service_id = null, $ret = array())
+function updateServiceHost($service_id = null, $ret = array(), $from_MC = false)
 {
     global $form, $pearDB;
     if (!$service_id) {
@@ -2009,9 +2011,26 @@ function updateServiceHost($service_id = null, $ret = array())
         }
     }
 
-    $rq = "DELETE FROM host_service_relation ";
-    $rq .= "WHERE service_service_id = '".$service_id."'";
-    $DBRESULT = $pearDB->query($rq);
+    if (!$from_MC) {
+        $rq = "DELETE FROM host_service_relation "
+            . "WHERE service_service_id = '" . $service_id . "' ";
+        $DBRESULT = $pearDB->query($rq);
+    } else {
+        # Purge service to host relations
+        if (count($ret1)) {
+            $rq = "DELETE FROM host_service_relation "
+                . "WHERE service_service_id = '" . $service_id . "' "
+                . "AND host_host_id IS NOT NULL ";
+            $DBRESULT = $pearDB->query($rq);
+        }
+        # Purge service to hostgroup relations
+        if (count($ret2)) {
+            $rq = "DELETE FROM host_service_relation "
+                . "WHERE service_service_id = '" . $service_id . "' "
+                . "AND hostgroup_hg_id IS NOT NULL ";
+            $DBRESULT = $pearDB->query($rq);
+        }
+    }
 
     if (count($ret2)) {
         for ($i = 0; $i < count($ret2); $i++)	{
